refactor: split doctor and hook_runtime hotspots
Behavior-preserving extraction of command-audit reliability, event coverage,
config inspection, and report formatting from doctor.go, and transcript,
session, subagent, state, resolve, and audit-suppression helpers from
hook_runtime.go. Core files stay under the project size guideline.

Closes #1259
